feat(voice-call): integrate dynamic agent resolution into response-generator
Merges the dynamic agent-resolution logic from patches/response-generator.ts
directly into the real response-generator.ts. Refines fallback logic to
distinguish between no binding configured (keep main) vs binding configured
without agentId (fall back to default agent). Deletes the misplaced patch file.
